Crate base

Re-exports

pub use sys::platform;

pub use external_mapping::Error as ExternalMappingError;

pub use external_mapping::Result as ExternalMappingResult;

pub use external_mapping::*;

pub use scoped_event_macro::*;

pub use sys::unix;

pub use unix::net::*;

pub use unix::ioctl::*;

pub use crate::descriptor::AsRawDescriptor;

pub use crate::descriptor::AsRawDescriptors;

pub use crate::descriptor::Descriptor;

pub use crate::descriptor::FromRawDescriptor;

pub use crate::descriptor::IntoRawDescriptor;

pub use crate::descriptor::SafeDescriptor;

pub use platform::*;

Modules

alloc

descriptor

descriptor_reflection

Provides infrastructure for de/serializing descriptors embedded in Rust data structures.

errno

event

external_mapping

ioctl

mmap

notifiers

scoped_event_macro

shm

sys

timer

tube

wait_context

Macros

debug

A macro for logging debug information.

error

A macro for logging an error.

generate_scoped_event

handle_eintr

Macro that retries the given expression every time its result indicates it was interrupted (i.e. returned EINTR). This is useful for operations that are prone to being interrupted by signals, such as blocking syscalls.

handle_eintr_errno

Macro that retries the given expression every time its result indicates it was interrupted. It is intended to use with system functions that signal error by returning -1 and setting errno to appropriate error code (EINTR, EINVAL, etc.) Most of standard non-reentrant libc functions use this way of signalling errors.

handle_eintr_rc

Macro that retries the given expression every time its result indicates it was interrupted. It is intended to use with system functions that return EINTR and other error codes directly as their result. Most of reentrant functions use this way of signalling errors.

info

A macro for logging info.

ioctl_expr

Raw macro to declare the expression that calculates an ioctl number

ioctl_io_nr

Declare an ioctl that transfers no data.

ioctl_ioc_nr

Raw macro to declare a function that returns an ioctl number.

ioctl_ior_nr

Declare an ioctl that reads data.

ioctl_iow_nr

Declare an ioctl that writes data.

ioctl_iowr_nr

Declare an ioctl that reads and writes data.

log

A macro for logging at an arbitrary priority level.

syscall

volatile_at_impl

volatile_impl

warn

A macro for logging a warning.

Structs

Error

A system error In Unix systems, retrieved from errno (man 3 errno), set by a libc function that returned an error. On Windows, retrieved from GetLastError, set by a Windows function that returned an error

Event

See EventFd for struct- and method-level documentation.

FakeTimer

See FakeTimerFd for struct- and method-level documentation.

LayoutAllocation

A contiguous memory allocation with a specified size and alignment, with a Drop impl to perform the deallocation.

MemoryMapping

See MemoryMapping for struct- and method-level documentation.

MemoryMappingBuilder

RecvTube

A Tube end which can only recv messages.

ScopedEvent

An Event wrapper which triggers when it goes out of scope.

SendTube

A Tube end which can only send messages. Cloneable.

SharedMemory

See SharedMemory for struct- and method-level documentation.

Timer

See TimerFd for struct- and method-level documentation.

TriggeredEvent

Represents an event that has been signaled and waited for via a wait function.

Tube

Bidirectional tube that support both send and recv.

WaitContext

Used to wait for multiple objects which are eligible for waiting.

Enums

EventReadResult

Wrapper around the return value of doing a read on an EventFd which distinguishes between getting a valid count of the number of times the eventfd has been written to and timing out waiting for the count to be non-zero.

EventType

Represents types of events to watch for.

TubeError

Traits

CloseNotifier

EventToken

MemoryMappingBuilderUnix

MemoryMappingUnix

ReadNotifier

SharedMemoryUnix

Functions

errno_result

Returns the last errno as a Result that is always an error.

ioctl⚠

Run an ioctl with no arguments.

ioctl_with_mut_ptr⚠

Run an ioctl with a mutable raw pointer.

ioctl_with_mut_ref⚠

Run an ioctl with a mutable reference.

ioctl_with_ptr⚠

Run an ioctl with a raw pointer.

ioctl_with_ref⚠

Run an ioctl with an immutable reference.

ioctl_with_val⚠

Run an ioctl with a single value argument.

Type Definitions

Result

TubeResult