Module base::common

Small system utility modules for usage by other higher level system utility modules like sys_util(on unix/linux) and win_sys_util(on windows).

Crates other than sys_util and win_sys_util should not depend directly on sys_util_core.

sys_util_core contains system utilities that are strictly platform/os independent. Platform dependent, conditionally compiled, code should not be added to sys_util_core.

Macros

generate_scoped_event

Structs

Error

A system error In Unix systems, retrieved from errno (man 3 errno), set by a libc function that returned an error. On Windows, retrieved from GetLastError, set by a Windows function that returned an error

ExternalMapping

ExternalMapping wraps an external library mapping. This is useful in cases where where the device memory is not compatible with the mmap interface, such as Vulkan VkDeviceMemory in the non-exportable case or when exported as an opaque fd.

LayoutAllocation

A contiguous memory allocation with a specified size and alignment, with a Drop impl to perform the deallocation.

Enums

Error

ExternalMappingError

Functions

errno_result

Returns the last errno as a Result that is always an error.

Type Definitions

ExternalMappingResult

Map

Result

Result

Unmap