Module base::error_chain

Re-exports from error_chain.

Modules

example_generated

These modules show an example of code generated by the macro. IT MUST NOT BE USED OUTSIDE THIS CRATE.

Macros

bail

Exits a function early with an error

ensure

Exits a function early with an error if the condition is not satisfied

error_chain

Macro for generating error types and traits. See crate level documentation for details.

quick_main

Convenient wrapper to be able to use ? and such in the main. You can use it with a separated function:

stringify_internal

From https://github.com/tailhook/quick-error Changes:

write_internal

Macro used interally for output expanding an expression

Structs

Backtrace

Representation of an owned and self-contained backtrace.

DisplayChain

A struct which formats an error for output.

Iter

Iterator over the error chain using the Error::cause() method.

Traits

ChainedError

This trait is implemented on all the errors generated by the error_chain macro.

ExitCode

Represents a value that can be used as the exit status of the process. See quick_main!.