Trait base::file_traits::FileGetLen

pub trait FileGetLen {
    fn get_len(&self) -> Result<u64, Error>;
}

A trait for getting the size of a file. This is equivalent to File’s metadata().len() method, but wrapped in a trait so that it can be implemented for other types.

Required methods

fn get_len(&self) -> Result<u64, Error>

Get the current length of the file in bytes.