base::generic_channel

Module oneshot

Structs

GenericOneshotReceiver

The oneshot receiver struct

GenericOneshotSender

The oneshot sender struct

Functions

oneshot

Creates a oneshot generic channel. This channel allows only a fixed capacity and might have other optimizations. The send and receive methods will consume the Sender/Receiver. We will automatically select ipc or crossbeam channels.