base::generic_channel

Trait GenericSend

pub trait GenericSend<T>
where
    T: Serialize + for<'de> Deserialize<'de>,
{
    // Required methods
    fn send(&self, _: T) -> SendResult;
    fn sender(&self) -> GenericSender<T>;
}

Abstraction of the ability to send a particular type of message cross-process. This can be used to ease the use of GenericSender sub-fields.

Required Methods

fn send(&self, _: T) -> SendResult

send message T

fn sender(&self) -> GenericSender<T>

get underlying sender