base

Module id

Namespaces and ids shared by many crates in Servo.

Macros

namespace_id 🔒

size_of_test 🔒

Asserts the size of a type at compile time.

Structs

BlobIndex

BroadcastChannelRouterIndex

BrowsingContextGroupId

BrowsingContextIndex

CookieStoreIndex

DomExceptionIndex

DomMatrixIndex

DomPointIndex

DomQuadIndex

DomRectIndex

HistoryStateIndex

ImageBitmapIndex

ImageDataIndex

Index

A non-zero index, associated with a particular type.

MessagePortIndex

MessagePortRouterIndex

NamespaceIndex

A pipeline-namespaced index associated with a particular type.

OffscreenCanvasIndex

PainterId

PipelineIndex

PipelineNamespace

Each pipeline ID needs to be unique. However, it also needs to be possible to generate the pipeline ID from an iframe element (this simplifies a lot of other code that makes use of pipeline IDs).

PipelineNamespaceId

PipelineNamespaceInstaller

A per-process installer of pipeline-namespaces.

PipelineNamespaceRequest

Request a pipeline-namespace id from the constellation.

QuotaExceededErrorIndex

ScriptEventLoopId

ScrollTreeNodeId

An id for a ScrollTreeNode in the ScrollTree. This contains both the index to the node in the tree’s array of nodes as well as the corresponding SpatialId for the SpatialNode in the WebRender display list.

ServiceWorkerIndex

ServiceWorkerRegistrationIndex

WebViewId

ZeroIndex

An attempt to create a new Index value failed because the index value was zero.

Constants

INSTALLED_SCRIPT_EVENT_LOOP_ID

PIPELINE_NAMESPACE

TEST_BROWSING_CONTEXT_ID

TEST_BROWSING_CONTEXT_INDEX

TEST_NAMESPACE

TEST_PAINTER_ID

TEST_PIPELINE_ID

TEST_PIPELINE_INDEX

TEST_SCRIPT_EVENT_LOOP_ID

TEST_WEBVIEW_ID

Statics

PAINTER_ID 🔒

PIPELINE_NAMESPACE_INSTALLER 🔒

A per-process unique pipeline-namespace-installer. Accessible via PipelineNamespace.

SCRIPT_EVENT_LOOP_ID 🔒

Traits

Indexable

A type that implements this trait is expected to be used as part of the NamespaceIndex type.

Type Aliases

BlobId

BroadcastChannelRouterId

BrowsingContextId

CookieStoreId

DomExceptionId

DomMatrixId

DomPointId

DomQuadId

DomRectId

HistoryStateId

ImageBitmapId

ImageDataId

MessagePortId

MessagePortRouterId

OffscreenCanvasId

PipelineId

QuotaExceededErrorId

ServiceWorkerId

ServiceWorkerRegistrationId