Struct base::id::PipelineNamespaceInstaller

pub struct PipelineNamespaceInstaller {
    request_sender: Option<IpcSender<PipelineNamespaceRequest>>,
    namespace_sender: IpcSender<PipelineNamespaceId>,
    namespace_receiver: IpcReceiver<PipelineNamespaceId>,
}

A per-process installer of pipeline-namespaces.

Fields
request_sender: Option<IpcSender<PipelineNamespaceRequest>>namespace_sender: IpcSender<PipelineNamespaceId>namespace_receiver: IpcReceiver<PipelineNamespaceId>
Implementations

impl PipelineNamespaceInstaller

pub fn set_sender(&mut self, sender: IpcSender<PipelineNamespaceRequest>)

Provide a request sender to send requests to the constellation.

pub fn install_namespace(&self)

Install a namespace, requesting a new Id from the constellation.