base::id

Trait Indexable

pub trait Indexable {
    const DISPLAY_PREFIX: &'static str;
}

A type that implements this trait is expected to be used as part of the NamespaceIndex type.

Required Associated Constants

const DISPLAY_PREFIX: &'static str

The string prefix to display when debug printing an instance of this type.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.