Module base::math

This module reexports everything from the cgmath crate and defines a few own types, such as AxialVector and AxialPoint.

Reexports

pub use self::cgmath::*;

Modules

billboard

Structs

AxialPoint
A 2-dimensional point in axial coordinates. See here for more information.

AxialVector
A 2-dimensional vector in axial coordinates. See here for more information.

Dimension2
A two-dimensional dimension.

Constants

SQRT_3

Functions

random_vec_with_angle
Generates a random Vector3f that has a specific angle to v.

Type Definitions

AxialType

DefaultFloat

DefaultInt

DefaultUnsigned

Dimension2f

Dimension2i

Dimension2u

Point2f

Point2i

Point2u

Point3f

Point3i

Point3u

Vector2f

Vector2i

Vector2u

Vector3f

Vector3i

Vector3u

Vector4f

Vector4i

Vector4u