Struct base::math::AxialPoint

pub struct AxialPoint {
    pub q: AxialType,
    pub r: AxialType,
}

A 2-dimensional point in axial coordinates. See here for more information.

Fields
q: AxialType r: AxialType
Methods

impl AxialPoint

fn new(q: AxialType, r: AxialType) -> Self

fn to_real(&self) -> Point2f

Returns the position of the hexagons center in the standard coordinate system using world::{HEX_INNER_RADIUS, HEX_OUTER_RADIUS}.

fn from_real(real: Point2f) -> Self

Return the AxialPoint from a Point2f. It's the pixel to hex algorithm from redblobgames.com.

fn s(&self) -> AxialType

Returns the s component of corresponding cube coordinates. In cube coordinates 'q + r + s = 0', so saving s is redundant and can be calculated on the fly when needed.