Struct base::math::AxialVector

pub struct AxialVector {
    pub q: AxialType,
    pub r: AxialType,
}

A 2-dimensional vector in axial coordinates. See here for more information.

Fields
q: AxialType r: AxialType
Methods

impl AxialVector

AxialVector defines a vector specifically for Axial cordinate system.

fn new(q: AxialType, r: AxialType) -> Self

fn to_real(&self) -> Vector2f

Returns the position of the hexagons center in the standard coordinate system using world::{HEX_INNER_RADIUS, HEX_OUTER_RADIUS}.

fn s(&self) -> AxialType

Returns the s component of corresponding cube coordinates. In cube coordinates 'q + r + s = 0', so saving s is redundant and can be calculated on the fly when needed.

fn unit_q() -> AxialVector

unit_q creates an default AxialVector with q:1 r:0.

fn unit_r() -> AxialVector

unit_r creates an default AxialVector with q:0 r:1.