Struct base::net::UnixSeqpacket

pub struct UnixSeqpacket { /* fields omitted */ }

A Unix SOCK_SEQPACKET socket point to given path

Implementations

impl UnixSeqpacket

pub fn connect<P>(path: P) -> Result<UnixSeqpacket, Error> where
P: AsRef<Path>,

Open a SOCK_SEQPACKET connection to socket named by path.

Arguments

path - Path to SOCK_SEQPACKET socket

Returns

A UnixSeqpacket structure point to the socket

Errors

Return io::Error when error occurs.

pub fn pair() -> Result<(UnixSeqpacket, UnixSeqpacket), Error>

Creates a pair of connected SOCK_SEQPACKET sockets.

Both returned file descriptors have the CLOEXEC flag set.s

pub fn try_clone(&self) -> Result<UnixSeqpacket, Error>

Clone the underlying FD.

pub fn get_readable_bytes(&self) -> Result<usize, Error>

Gets the number of bytes that can be read from this socket without blocking.

pub fn next_packet_size(&self) -> Result<usize, Error>

Gets the number of bytes in the next packet. This blocks as if recv were called, respecting the blocking and timeout settings of the underlying socket.

pub fn send(&self, buf: &[u8]) -> Result<usize, Error>

Write data from a given buffer to the socket fd

Arguments

buf - A reference to the data buffer.

Returns

usize - The size of bytes written to the buffer.

Errors

Returns error when libc::write failed.

pub fn recv(&self, buf: &mut [u8]) -> Result<usize, Error>

Read data from the socket fd to a given buffer

Arguments

buf - A mut reference to the data buffer.

Returns

usize - The size of bytes read to the buffer.

Errors

Returns error when libc::read failed.

pub fn recv_to_vec(&self, buf: &mut Vec<u8, Global>) -> Result<(), Error>

Read data from the socket fd to a given Vec, resizing it to the received packet’s size.

Arguments

buf - A mut reference to a Vec to resize and read into.

Errors

Returns error when libc::read or get_readable_bytes failed.

pub fn recv_as_vec(&self) -> Result<Vec<u8, Global>, Error>

Read data from the socket fd to a new Vec.

Returns

vec - A new Vec with the entire received packet.

Errors

Returns error when libc::read or get_readable_bytes failed.

pub fn recv_as_vec_with_fds(
&self
) -> Result<(Vec<u8, Global>, Vec<i32, Global>), Error>

Read data and fds from the socket fd to a new pair of Vec.

Returns

Vec<u8> - A new Vec with the entire received packet’s bytes.

Vec<RawFd> - A new Vec with the entire received packet’s fds.

Errors

Returns error when recv_with_fds or get_readable_bytes failed.

pub fn set_read_timeout(&self, timeout: Option<Duration>) -> Result<(), Error>

Sets or removes the timeout for read/recv operations on this socket.

pub fn set_write_timeout(&self, timeout: Option<Duration>) -> Result<(), Error>

Sets or removes the timeout for write/send operations on this socket.