Module base::prelude

Common imports.

Re-exports

pub use crate::convert;

pub use crate::destroy;

pub use crate::err;

pub use crate::err::Res;

pub use crate::err::ResExt;

pub use crate::error_chain;

pub use crate::identity;

pub use crate::time;

pub use crate::time::DurationExt;

pub use crate::time_stats;

pub use crate::uid;

pub use crate::Range;

pub use crate::SVec16;

pub use crate::SVec32;

pub use crate::SVec64;

pub use crate::SVec8;

pub use crate::SampleRate;

Modules

fmt

Utilities for formatting and printing Strings.

log

Log macros re-exports.

ops

Overloadable operators.

serde

Serde trait re-exports.

sync

Useful synchronization primitives.

Macros

bail

Exits a function early with an error

lazy_static

Structs

Arc

A thread-safe reference-counting pointer. ‘Arc’ stands for ‘Atomically Reference Counted’.

BTMap

An ordered map based on a B-Tree.

BTSet

An ordered set based on a B-Tree.

HMap

A hash map implemented with quadratic probing and SIMD lookup.

HSet

A hash set implemented as a HashMap where the value is ().

Enums

Either

The enum Either with variants Left and Right is a general purpose sum type with two cases.

Inhabited

Inhabited type.

Traits

Borrow

A trait for borrowing data.

FromStr

Parse a value from a string

TryFrom

Simple and safe type conversions that may fail in a controlled way under some circumstances. It is the reciprocal of TryInto.

TryInto

An attempted conversion that consumes self, which may or may not be expensive.