Struct base::print_tree::PrintTree

pub struct PrintTree {
    level: u32,
    queued_item: Option<String>,
}

A struct that makes it easier to print out a pretty tree of data, which can be visually scanned more easily.

Fields
level: u32

The current level of recursion.

queued_item: Option<String>

An item which is queued up, so that we can determine if we need a mid-tree prefix or a branch ending prefix.

Implementations

impl PrintTree

pub fn new(title: String) -> PrintTree

pub fn new_level(&mut self, queued_title: String)

Descend one level in the tree with the given title string.

pub fn end_level(&mut self)

Ascend one level in the tree.

pub fn add_item(&mut self, text: String)

Add an item to the current level in the tree.

fn print_level_prefix(&self)

fn print_level_child_indentation(&self)

fn flush_queued_item(&mut self, prefix: &str)