Trait base::signal::Killable

pub unsafe trait Killable {
    pub fn pthread_handle(&self) -> u64;

    pub fn kill(&self, num: i32) -> Result<(), Error> { ... }
}

Trait for threads that can be signalled via pthread_kill.

Note that this is only useful for signals between SIGRTMIN and SIGRTMAX because these are guaranteed to not be used by the C runtime.

This is marked unsafe because the implementation of this trait must guarantee that the returned pthread_t is valid and has a lifetime at least that of the trait object.

Required methods

pub fn pthread_handle(&self) -> u64

Loading content...
Provided methods

pub fn kill(&self, num: i32) -> Result<(), Error>

Sends the signal num to this killable thread.

The value of num must be within [SIGRTMIN, SIGRTMAX] range.

Loading content...