Struct base::EventFd

pub struct EventFd { /* fields omitted */ }

A safe wrapper around a Linux eventfd (man 2 eventfd).

An eventfd is useful because it is sendable across processes and can be used for signaling in and out of the KVM API. They can also be polled like any other file descriptor.

Implementations

impl EventFd

pub fn new() -> Result<EventFd, Error>

Creates a new blocking EventFd with an initial value of 0.

pub fn write(&self, v: u64) -> Result<(), Error>

Adds v to the eventfd's count, blocking until this won't overflow the count.

pub fn read(&self) -> Result<u64, Error>

Blocks until the the eventfd's count is non-zero, then resets the count to zero.

pub fn read_timeout(
&mut self,
timeout: Duration
) -> Result<EventReadResult, Error>

Blocks for a maximum of timeout duration until the the eventfd's count is non-zero. If a timeout does not occur then the count is returned as a EventReadResult::Count(count), and the count is reset to 0. If a timeout does occur then this function will return EventReadResult::Timeout.

pub fn try_clone(&self) -> Result<EventFd, Error>

Clones this EventFd, internally creating a new file descriptor. The new EventFd will share the same underlying count within the kernel.