Struct base::MemoryMapping

pub struct MemoryMapping { /* fields omitted */ }

Wraps an anonymous shared memory mapping in the current process.

Implementations

impl MemoryMapping

pub fn new(size: usize) -> Result<MemoryMapping, Error>

Creates an anonymous shared, read/write mapping of size bytes.

Arguments

size - Size of memory region in bytes.

pub fn new_protection(
size: usize,
prot: Protection
) -> Result<MemoryMapping, Error>

Creates an anonymous shared mapping of size bytes with prot protection.

Arguments

size - Size of memory region in bytes.

prot - Protection (e.g. readable/writable) of the memory region.

pub fn from_fd(fd: &dyn AsRawFd, size: usize) -> Result<MemoryMapping, Error>

Maps the first size bytes of the given fd as read/write.

Arguments

fd - File descriptor to mmap from.

size - Size of memory region in bytes.

pub fn from_fd_offset(
fd: &dyn AsRawFd,
size: usize,
offset: u64
) -> Result<MemoryMapping, Error>

pub fn from_fd_offset_protection(
fd: &dyn AsRawFd,
size: usize,
offset: u64,
prot: Protection
) -> Result<MemoryMapping, Error>

Maps the size bytes starting at offset bytes of the given fd as read/write.

Arguments

fd - File descriptor to mmap from.

size - Size of memory region in bytes.

offset - Offset in bytes from the beginning of fd to start the mmap.

prot - Protection (e.g. readable/writable) of the memory region.

pub fn from_fd_offset_protection_populate(
fd: &dyn AsRawFd,
size: usize,
offset: u64,
prot: Protection,
populate: bool
) -> Result<MemoryMapping, Error>

Maps size bytes starting at offset from the given fd as read/write, and requests that the pages are pre-populated.

Arguments

fd - File descriptor to mmap from.

size - Size of memory region in bytes.

offset - Offset in bytes from the beginning of fd to start the mmap.

pub unsafe fn new_protection_fixed(
addr: *mut u8,
size: usize,
prot: Protection
) -> Result<MemoryMapping, Error>

Creates an anonymous shared mapping of size bytes with prot protection.

Arguments

addr - Memory address to mmap at.

size - Size of memory region in bytes.

prot - Protection (e.g. readable/writable) of the memory region.

Safety

This function should not be called before the caller unmaps any mmap'd regions already present at (addr..addr+size).

pub unsafe fn from_fd_offset_protection_fixed(
addr: *mut u8,
fd: &dyn AsRawFd,
size: usize,
offset: u64,
prot: Protection
) -> Result<MemoryMapping, Error>

Maps the size bytes starting at offset bytes of the given fd with prot protections.

Arguments

addr - Memory address to mmap at.

fd - File descriptor to mmap from.

size - Size of memory region in bytes.

offset - Offset in bytes from the beginning of fd to start the mmap.

prot - Protection (e.g. readable/writable) of the memory region.

Safety

This function should not be called before the caller unmaps any mmap'd regions already present at (addr..addr+size).

pub fn use_hugepages(&self) -> Result<(), Error>

Madvise the kernel to use Huge Pages for this mapping.

pub fn msync(&self) -> Result<(), Error>

Calls msync with MS_SYNC on the mapping.

pub fn write_slice(&self, buf: &[u8], offset: usize) -> Result<usize, Error>

Writes a slice to the memory region at the specified offset. Returns the number of bytes written. The number of bytes written can be less than the length of the slice if there isn't enough room in the memory region.

Examples

Write a slice at offset 256.

    let res = mem_map.write_slice(&[1,2,3,4,5], 256);
    assert!(res.is_ok());
    assert_eq!(res.unwrap(), 5);

pub fn read_slice(&self, buf: &mut [u8], offset: usize) -> Result<usize, Error>

Reads to a slice from the memory region at the specified offset. Returns the number of bytes read. The number of bytes read can be less than the length of the slice if there isn't enough room in the memory region.

Examples

Read a slice of size 16 at offset 256.

    let buf = &mut [0u8; 16];
    let res = mem_map.read_slice(buf, 256);
    assert!(res.is_ok());
    assert_eq!(res.unwrap(), 16);

pub fn write_obj<T>(&self, val: T, offset: usize) -> Result<(), Error> where
T: DataInit,

Writes an object to the memory region at the specified offset. Returns Ok(()) if the object fits, or Err if it extends past the end.

Examples

Write a u64 at offset 16.

    let res = mem_map.write_obj(55u64, 16);
    assert!(res.is_ok());

pub fn read_obj<T>(&self, offset: usize) -> Result<T, Error> where
T: DataInit,

Reads on object from the memory region at the given offset. Reading from a volatile area isn't strictly safe as it could change mid-read. However, as long as the type T is plain old data and can handle random initialization, everything will be OK.

Examples

Read a u64 written to offset 32.

    let res = mem_map.write_obj(55u64, 32);
    assert!(res.is_ok());
    let num: u64 = mem_map.read_obj(32).unwrap();
    assert_eq!(55, num);

pub fn read_to_memory(
&self,
mem_offset: usize,
src: &dyn AsRawFd,
count: usize
) -> Result<(), Error>

Reads data from a file descriptor and writes it to guest memory.

Arguments

mem_offset - Begin writing memory at this offset.

src - Read from src to memory.

count - Read count bytes from src to memory.

Examples

Read bytes from /dev/urandom

      let mut file = File::open(Path::new("/dev/urandom")).map_err(|_| ())?;
      mem_map.read_to_memory(32, &mut file, 128).map_err(|_| ())?;
      let rand_val: u32 =  mem_map.read_obj(40).map_err(|_| ())?;

pub fn write_from_memory(
&self,
mem_offset: usize,
dst: &dyn AsRawFd,
count: usize
) -> Result<(), Error>

Writes data from memory to a file descriptor.

Arguments

mem_offset - Begin reading memory from this offset.

dst - Write from memory to dst.

count - Read count bytes from memory to src.

Examples

Write 128 bytes to /dev/null

      let mut file = File::open(Path::new("/dev/null")).map_err(|_| ())?;
      mem_map.write_from_memory(32, &mut file, 128).map_err(|_| ())?;

pub fn remove_range(&self, mem_offset: usize, count: usize) -> Result<(), Error>

Uses madvise to tell the kernel to remove the specified range. Subsequent reads to the pages in the range will return zero bytes.