Struct base::MemoryMappingArena

pub struct MemoryMappingArena { /* fields omitted */ }

Tracks Fixed Memory Maps within an anonymous memory-mapped fixed-sized arena in the current process.

Implementations

impl MemoryMappingArena

pub fn new(size: usize) -> Result<MemoryMappingArena, Error>

Creates an mmap arena of size bytes.

Arguments

size - Size of memory region in bytes.

pub fn add_anon_protection(
&mut self,
offset: usize,
size: usize,
prot: Protection
) -> Result<(), Error>

Anonymously maps size bytes at offset bytes from the start of the arena with prot protections. offset must be page aligned.

Arguments

offset - Page aligned offset into the arena in bytes.

size - Size of memory region in bytes.

prot - Protection (e.g. readable/writable) of the memory region.

pub fn add_anon(&mut self, offset: usize, size: usize) -> Result<(), Error>

Anonymously maps size bytes at offset bytes from the start of the arena. offset must be page aligned.

Arguments

offset - Page aligned offset into the arena in bytes.

size - Size of memory region in bytes.

pub fn add_fd(
&mut self,
offset: usize,
size: usize,
fd: &dyn AsRawFd
) -> Result<(), Error>

Maps size bytes from the start of the given fd at offset bytes from the start of the arena. offset must be page aligned.

Arguments

offset - Page aligned offset into the arena in bytes.

size - Size of memory region in bytes.

fd - File descriptor to mmap from.

pub fn add_fd_offset(
&mut self,
offset: usize,
size: usize,
fd: &dyn AsRawFd,
fd_offset: u64
) -> Result<(), Error>

Maps size bytes starting at fs_offset bytes from within the given fd at offset bytes from the start of the arena. offset must be page aligned.

Arguments

offset - Page aligned offset into the arena in bytes.

size - Size of memory region in bytes.

fd - File descriptor to mmap from.

fd_offset - Offset in bytes from the beginning of fd to start the mmap.

pub fn add_fd_offset_protection(
&mut self,
offset: usize,
size: usize,
fd: &dyn AsRawFd,
fd_offset: u64,
prot: Protection
) -> Result<(), Error>

Maps size bytes starting at fs_offset bytes from within the given fd at offset bytes from the start of the arena with prot protections. offset must be page aligned.

Arguments

offset - Page aligned offset into the arena in bytes.

size - Size of memory region in bytes.

fd - File descriptor to mmap from.

fd_offset - Offset in bytes from the beginning of fd to start the mmap.

prot - Protection (e.g. readable/writable) of the memory region.

pub fn remove(&mut self, offset: usize, size: usize) -> Result<(), Error>

Removes size bytes at offset bytes from the start of the arena. offset must be page aligned.

Arguments

offset - Page aligned offset into the arena in bytes.

size - Size of memory region in bytes.