Struct base::Timer

pub struct Timer {
    pub(crate) handle: SafeDescriptor,
    pub(crate) interval: Option<Duration>,
}

Fields
handle: SafeDescriptorinterval: Option<Duration>
Implementations

impl Timer

pub fn try_clone(&self) -> Result<Timer, Error>

Creates a new Timer instance that shares the same underlying SafeDescriptor as the existing Timer instance.

impl Timer

pub fn new() -> Result<Timer>

Creates a new timerfd. The timer is initally disarmed and must be armed by calling reset.

fn set_time(
&mut self,
dur: Option<Duration>,
interval: Option<Duration>
) -> Result<()>

pub fn reset(&mut self, dur: Duration, interval: Option<Duration>) -> Result<()>

Sets the timer to expire after dur. If interval is not None it represents the period for repeated expirations after the initial expiration. Otherwise the timer will expire just once. Cancels any existing duration and repeating interval.

pub fn clear(&mut self) -> Result<()>

Disarms the timer.

pub fn wait_for(&mut self, timeout: Option<Duration>) -> Result<WaitResult>

Waits until the timer expires or an optional wait timeout expires, whichever happens first.

Returns

WaitResult::Expired if the timer expired.

WaitResult::Timeout if timeout was not None and the timer did not expire within the specified timeout period.

pub fn wait(&mut self) -> Result<WaitResult>

Waits until the timer expires.

pub fn mark_waited(&mut self) -> Result<bool>

After a timer is triggered from an EventContext, mark the timer as having been waited for. If a timer is not marked waited, it will immediately trigger the event context again. This does not need to be called after calling Timer::wait.

Returns true if the timer has been adjusted since the EventContext was triggered by this timer.

pub fn resolution() -> Result<Duration>

Returns the resolution of timers on the host.