Module base::sys::unix::descriptor

Macros

AsRawDescriptor

FromRawDescriptor

IntoRawDescriptor

Constants

INVALID_DESCRIPTOR

KCMP_FILE

Functions

clear_descriptor_cloexec

Clears CLOEXEC flag on descriptor

clear_fd_cloexec

Clears CLOEXEC flag on fd

clone_descriptor

Clones descriptor, returning a new RawDescriptor that refers to the same open file description as descriptor. The cloned descriptor will have the FD_CLOEXEC flag set but will not share any other file descriptor flags with descriptor.

clone_fd

Clones fd, returning a new file descriptor that refers to the same open file description as fd. The cloned fd will have the FD_CLOEXEC flag set but will not share any other file descriptor flags with fd.

Type Definitions

RawDescriptor