Module base::sys::unix::file_traits::lib

Structs

IoBufMut

This type is essentialy std::io::IoBufMut, and guaranteed to be ABI-compatible with libc::iovec; however, it does NOT automatically deref to &mut [u8], which is critical because it can point to guest memory. (Guest memory is implicitly mutably borrowed by the guest, so another mutable borrow would violate Rust assumptions about references.)

VolatileSlice

A slice of raw memory that supports volatile access. Like std::io::IoBufMut, this type is guaranteed to be ABI-compatible with libc::iovec but unlike IoBufMut, it doesn’t automatically deref to &mut [u8].

iovec

Enums

c_void

Equivalent to C’s void type when used as a pointer.

Functions

pread64⚠

preadv64⚠

pwrite64⚠

pwritev64⚠

read⚠

readv⚠

write⚠

writev⚠

Type Definitions

c_int

off64_t

size_t