Module base::sys::unix::mmap

The mmap module provides a safe interface to mmap memory and ensures unmap is called when the mmap object leaves scope.

Structs

MemoryMapping

Wraps an anonymous shared memory mapping in the current process. Provides RAII semantics including munmap when no longer needed.

MemoryMappingArena

Tracks Fixed Memory Maps within an anonymous memory-mapped fixed-sized arena in the current process.

Protection

Memory access type for anonymous shared memory mapping.

Enums

Error

Traits

MappedRegion

A range of memory that can be msynced, for abstracting over different types of memory mappings.

MemoryMappingBuilderUnix

Unix

Functions

validate_includes_range

Validates that offset..offset+range_size lies within the bounds of a memory mapping of mmap_size bytes. Also checks for any overflow.

Type Definitions

Result