Struct base::sys::unix::mmap::MemoryMapping

pub struct MemoryMapping {
    addr: *mut u8,
    size: usize,
}

Wraps an anonymous shared memory mapping in the current process. Provides RAII semantics including munmap when no longer needed.

Fields
addr: *mut u8size: usize
Implementations

impl MemoryMapping

pub fn new(size: usize) -> Result<MemoryMapping, Error>

Creates an anonymous shared, read/write mapping of size bytes.

Arguments

size - Size of memory region in bytes.

pub fn new_protection(
size: usize,
prot: Protection
) -> Result<MemoryMapping, Error>

Creates an anonymous shared mapping of size bytes with prot protection.

Arguments

size - Size of memory region in bytes.

prot - Protection (e.g. readable/writable) of the memory region.

pub fn from_fd(
fd: &dyn AsRawDescriptor,
size: usize
) -> Result<MemoryMapping, Error>

Maps the first size bytes of the given fd as read/write.

Arguments

fd - File descriptor to mmap from.

size - Size of memory region in bytes.

pub fn from_fd_offset(
fd: &dyn AsRawDescriptor,
size: usize,
offset: u64
) -> Result<MemoryMapping, Error>

fn from_fd_offset_flags(
fd: &dyn AsRawDescriptor,
size: usize,
offset: u64,
flags: c_int,
prot: Protection
) -> Result<MemoryMapping, Error>

Maps the size bytes starting at offset bytes of the given fd as read/write.

Arguments

fd - File descriptor to mmap from.

size - Size of memory region in bytes.

offset - Offset in bytes from the beginning of fd to start the mmap.

flags - flags passed directly to mmap.

prot - Protection (e.g. readable/writable) of the memory region.

pub fn from_fd_offset_protection(
fd: &dyn AsRawDescriptor,
size: usize,
offset: u64,
prot: Protection
) -> Result<MemoryMapping, Error>

Maps the size bytes starting at offset bytes of the given fd as read/write.

Arguments

fd - File descriptor to mmap from.

size - Size of memory region in bytes.

offset - Offset in bytes from the beginning of fd to start the mmap.

prot - Protection (e.g. readable/writable) of the memory region.

pub fn from_fd_offset_protection_populate(
fd: &dyn AsRawDescriptor,
size: usize,
offset: u64,
prot: Protection,
populate: bool
) -> Result<MemoryMapping, Error>

Maps size bytes starting at offset from the given fd as read/write, and requests that the pages are pre-populated.

Arguments

fd - File descriptor to mmap from.

size - Size of memory region in bytes.

offset - Offset in bytes from the beginning of fd to start the mmap.

pub unsafe fn new_protection_fixed(
addr: *mut u8,
size: usize,
prot: Protection
) -> Result<MemoryMapping, Error>

Creates an anonymous shared mapping of size bytes with prot protection.

Arguments

addr - Memory address to mmap at.

size - Size of memory region in bytes.

prot - Protection (e.g. readable/writable) of the memory region.

Safety

This function should not be called before the caller unmaps any mmap’d regions already present at (addr..addr+size).

pub unsafe fn from_descriptor_offset_protection_fixed(
addr: *mut u8,
fd: &dyn AsRawDescriptor,
size: usize,
offset: u64,
prot: Protection
) -> Result<MemoryMapping, Error>

Maps the size bytes starting at offset bytes of the given fd with prot protections.

Arguments

addr - Memory address to mmap at.

fd - File descriptor to mmap from.

size - Size of memory region in bytes.

offset - Offset in bytes from the beginning of fd to start the mmap.

prot - Protection (e.g. readable/writable) of the memory region.

Safety

This function should not be called before the caller unmaps any mmap’d regions already present at (addr..addr+size).

unsafe fn try_mmap(
addr: Option<*mut u8>,
size: usize,
prot: c_int,
flags: c_int,
fd: Option<(&dyn AsRawDescriptor, u64)>
) -> Result<MemoryMapping, Error>

Helper wrapper around libc::mmap that does some basic validation, and calls madvise with MADV_DONTDUMP on the created mmap

pub fn use_hugepages(&self) -> Result<(), Error>

Madvise the kernel to use Huge Pages for this mapping.

pub fn msync(&self) -> Result<(), Error>

Calls msync with MS_SYNC on the mapping.

pub fn read_to_memory(
&self,
mem_offset: usize,
src: &dyn AsRawDescriptor,
count: usize
) -> Result<(), Error>

Reads data from a file descriptor and writes it to guest memory.

Arguments

mem_offset - Begin writing memory at this offset.

src - Read from src to memory.

count - Read count bytes from src to memory.

Examples

Read bytes from /dev/urandom

      let mut file = File::open(Path::new("/dev/urandom")).map_err(|_| ())?;
      mem_map.read_to_memory(32, &mut file, 128).map_err(|_| ())?;
      let rand_val: u32 =  mem_map.read_obj(40).map_err(|_| ())?;

pub fn write_from_memory(
&self,
mem_offset: usize,
dst: &dyn AsRawDescriptor,
count: usize
) -> Result<(), Error>

Writes data from memory to a file descriptor.

Arguments

mem_offset - Begin reading memory from this offset.

dst - Write from memory to dst.

count - Read count bytes from memory to src.

Examples

Write 128 bytes to /dev/null

      let mut file = File::open(Path::new("/dev/null")).map_err(|_| ())?;
      mem_map.write_from_memory(32, &mut file, 128).map_err(|_| ())?;

pub fn remove_range(&self, mem_offset: usize, count: usize) -> Result<(), Error>

Uses madvise to tell the kernel to remove the specified range. Subsequent reads to the pages in the range will return zero bytes.

pub fn lock_all(&self) -> Result<(), Error>

Disable host swap for this mapping.

pub(crate) fn range_end(
&self,
offset: usize,
count: usize
) -> Result<usize, Error>