Module base::sys::unix::netlink

Macros

debug_pr

Structs

GenlMsgHdr

Generic netlink header struct, can be used by netlink consumer

NetlinkGenericDataIter

Iterator over struct NlAttr as received from a netlink socket.

NetlinkGenericRead

NetlinkGenericSocket

Safe wrapper for NETLINK_GENERIC netlink sockets.

NetlinkMessage

A single netlink message, including its header and data.

NetlinkMessageIter

Iterator over struct nlmsghdr as received from a netlink socket.

NlAttr

Netlink attribute struct, can be used by netlink consumer

NlAttrWithData

NlMsgHdr

Constants

GENL_HDRLEN

NLATTR_ALIGN_TO

NLA_HDRLEN

NLMSGHDR_SIZE

Functions

nlattr_align

parse_ctrl_group_name_and_id

parse_ctrl_mcast_group_id

Parse CTRL_ATTR_MCAST_GROUPS data in order to get multicast group id

strip_padding