Module base::sys::unix::sock_ctrl_msg

Used to send and receive messages with file descriptors on sockets that accept control messages (e.g. Unix domain sockets).

Macros

CMSG_ALIGN

CMSG_LEN

CMSG_SPACE

Enums

CmsgBuffer

Constants

CMSG_BUFFER_INLINE_CAPACITY

SCM_SOCKET_MAX_FD_COUNT

The maximum number of FDs that can be sent in a single send.

Traits

AsIobuf

Trait for types that can be converted into an iovec that can be referenced by a syscall for the lifetime of this object.

ScmSocket

Trait for file descriptors can send and receive socket control messages via sendmsg and recvmsg.

Functions

CMSG_DATA

get_next_cmsg

raw_recvmsg

raw_sendmsg