Trait base::sys::unix::sock_ctrl_msg::AsIobuf

pub unsafe trait AsIobuf: Sized {
    fn as_iobuf(&self) -> iovec;

    fn as_iobuf_slice(bufs: &[Self]) -> &[iovec];
}

Trait for types that can be converted into an iovec that can be referenced by a syscall for the lifetime of this object.

This trait is unsafe because interfaces that use this trait depend on the base pointer and size being accurate.

Required methods

fn as_iobuf(&self) -> iovec

Returns a iovec that describes a contiguous region of memory.

fn as_iobuf_slice(bufs: &[Self]) -> &[iovec]

Returns a slice of iovecs that each describe a contiguous region of memory.