Trait base::FileReadWriteVolatile

pub trait FileReadWriteVolatile {
    pub fn read_volatile(
        &mut self, 
        slice: VolatileSlice<'_>
    ) -> Result<usize, Error>;

    pub fn write_volatile(
        &mut self, 
        slice: VolatileSlice<'_>
    ) -> Result<usize, Error>;

    pub fn read_vectored_volatile(
        &mut self, 
        bufs: &[VolatileSlice<'_>]
    ) -> Result<usize, Error> { ... }

    pub fn read_exact_volatile(
        &mut self, 
        slice: VolatileSlice<'_>
    ) -> Result<(), Error> { ... }

    pub fn write_vectored_volatile(
        &mut self, 
        bufs: &[VolatileSlice<'_>]
    ) -> Result<usize, Error> { ... }

    pub fn write_all_volatile(
        &mut self, 
        slice: VolatileSlice<'_>
    ) -> Result<(), Error> { ... }
}

A trait similar to Read and Write, but uses volatile memory as buffers.

Required methods

pub fn read_volatile(
&mut self,
slice: VolatileSlice<'_>
) -> Result<usize, Error>

Read bytes from this file into the given slice, returning the number of bytes read on success.

pub fn write_volatile(
&mut self,
slice: VolatileSlice<'_>
) -> Result<usize, Error>

Write bytes from the slice to the given file, returning the number of bytes written on success.

Loading content...
Provided methods

pub fn read_vectored_volatile(
&mut self,
bufs: &[VolatileSlice<'_>]
) -> Result<usize, Error>

Like read_volatile, except it reads to a slice of buffers. Data is copied to fill each buffer in order, with the final buffer written to possibly being only partially filled. This method must behave as a single call to read_volatile with the buffers concatenated would. The default implementation calls read_volatile with either the first nonempty buffer provided, or returns Ok(0) if none exists.

pub fn read_exact_volatile(
&mut self,
slice: VolatileSlice<'_>
) -> Result<(), Error>

Reads bytes from this into the given slice until all bytes in the slice are written, or an error is returned.

pub fn write_vectored_volatile(
&mut self,
bufs: &[VolatileSlice<'_>]
) -> Result<usize, Error>

Like write_volatile, except that it writes from a slice of buffers. Data is copied from each buffer in order, with the final buffer read from possibly being only partially consumed. This method must behave as a call to write_volatile with the buffers concatenated would. The default implementation calls write_volatile with either the first nonempty buffer provided, or returns Ok(0) if none exists.

pub fn write_all_volatile(
&mut self,
slice: VolatileSlice<'_>
) -> Result<(), Error>

Write bytes from the slice to the given file until all the bytes from the slice have been written, or an error is returned.

Loading content...