Trait base::ReadNotifier

pub trait ReadNotifier {
    fn get_read_notifier(&self) -> &dyn AsRawDescriptor;
}

Required methods

fn get_read_notifier(&self) -> &dyn AsRawDescriptor

Gets a descriptor that can be used in EventContext to wait for events to be available (e.g. to avoid receive_events blocking).