Trait base::ScmSocket

pub trait ScmSocket {
    pub fn socket_fd(&self) -> i32;

    pub fn send_with_fd<D>(&self, buf: &[D], fd: i32) -> Result<usize, Error>
    where
        D: AsIobuf,
    { ... }

    pub fn send_with_fds<D>(
        &self, 
        buf: &[D], 
        fd: &[i32]
    ) -> Result<usize, Error>
    where
        D: AsIobuf,
    { ... }

    pub fn send_bufs_with_fd(
        &self, 
        bufs: &[&[u8]], 
        fd: i32
    ) -> Result<usize, Error> { ... }

    pub fn send_bufs_with_fds(
        &self, 
        bufs: &[&[u8]], 
        fd: &[i32]
    ) -> Result<usize, Error> { ... }

    pub fn recv_with_fd(
        &self, 
        buf: &mut [u8]
    ) -> Result<(usize, Option<File>), Error> { ... }

    pub fn recv_with_fds(
        &self, 
        buf: &mut [u8], 
        fds: &mut [i32]
    ) -> Result<(usize, usize), Error> { ... }

    pub fn recv_iovecs_with_fds(
        &self, 
        iovecs: &mut [iovec], 
        fds: &mut [i32]
    ) -> Result<(usize, usize), Error> { ... }
}

Trait for file descriptors can send and receive socket control messages via sendmsg and recvmsg.

Required methods

pub fn socket_fd(&self) -> i32

Gets the file descriptor of this socket.

Loading content...
Provided methods

pub fn send_with_fd<D>(&self, buf: &[D], fd: i32) -> Result<usize, Error> where
D: AsIobuf,

Sends the given data and file descriptor over the socket.

On success, returns the number of bytes sent.

Arguments

buf - A buffer of data to send on the socket.

fd - A file descriptors to be sent.

pub fn send_with_fds<D>(&self, buf: &[D], fd: &[i32]) -> Result<usize, Error> where
D: AsIobuf,

Sends the given data and file descriptors over the socket.

On success, returns the number of bytes sent.

Arguments

buf - A buffer of data to send on the socket.

fds - A list of file descriptors to be sent.

pub fn send_bufs_with_fd(&self, bufs: &[&[u8]], fd: i32) -> Result<usize, Error>

Sends the given data and file descriptor over the socket.

On success, returns the number of bytes sent.

Arguments

bufs - A slice of slices of data to send on the socket.

fd - A file descriptors to be sent.

pub fn send_bufs_with_fds(
&self,
bufs: &[&[u8]],
fd: &[i32]
) -> Result<usize, Error>

Sends the given data and file descriptors over the socket.

On success, returns the number of bytes sent.

Arguments

bufs - A slice of slices of data to send on the socket.

fds - A list of file descriptors to be sent.

pub fn recv_with_fd(
&self,
buf: &mut [u8]
) -> Result<(usize, Option<File>), Error>

Receives data and potentially a file descriptor from the socket.

On success, returns the number of bytes and an optional file descriptor.

Arguments

buf - A buffer to receive data from the socket.vm

pub fn recv_with_fds(
&self,
buf: &mut [u8],
fds: &mut [i32]
) -> Result<(usize, usize), Error>

Receives data and file descriptors from the socket.

On success, returns the number of bytes and file descriptors received as a tuple (bytes count, files count).

Arguments

buf - A buffer to receive data from the socket.

fds - A slice of RawFds to put the received file descriptors into. On success, the number of valid file descriptors is indicated by the second element of the returned tuple. The caller owns these file descriptors, but they will not be closed on drop like a File-like type would be. It is recommended that each valid file descriptor gets wrapped in a drop type that closes it after this returns.

pub fn recv_iovecs_with_fds(
&self,
iovecs: &mut [iovec],
fds: &mut [i32]
) -> Result<(usize, usize), Error>

Receives data and file descriptors from the socket.

On success, returns the number of bytes and file descriptors received as a tuple (bytes count, files count).

Arguments

ioves - A slice of iovecs to store received data.

fds - A slice of RawFds to put the received file descriptors into. On success, the number of valid file descriptors is indicated by the second element of the returned tuple. The caller owns these file descriptors, but they will not be closed on drop like a File-like type would be. It is recommended that each valid file descriptor gets wrapped in a drop type that closes it after this returns.

Loading content...