Trait base::Terminal

pub unsafe trait Terminal {
    pub fn tty_fd(&self) -> i32;

    pub fn set_canon_mode(&self) -> Result<(), Error> { ... }

    pub fn set_raw_mode(&self) -> Result<(), Error> { ... }

    pub fn set_non_block(&self, non_block: bool) -> Result<(), Error> { ... }
}

Trait for file descriptors that are TTYs, according to isatty(3).

This is marked unsafe because the implementation must promise that the returned RawFd is a valid fd and that the lifetime of the returned fd is at least that of the trait object.

Required methods

pub fn tty_fd(&self) -> i32

Gets the file descriptor of the TTY.

Loading content...
Provided methods

pub fn set_canon_mode(&self) -> Result<(), Error>

Set this terminal's mode to canonical mode (ICANON | ECHO | ISIG).

pub fn set_raw_mode(&self) -> Result<(), Error>

Set this terminal's mode to raw mode (!(ICANON | ECHO | ISIG)).

pub fn set_non_block(&self, non_block: bool) -> Result<(), Error>

Sets the non-blocking mode of this terminal's file descriptor.

If non_block is true, then read_raw will not block. If non_block is false, then read_raw may block if there is nothing to read.

Loading content...