Trait base::WriteZeroesAt

pub trait WriteZeroesAt {
    pub fn write_zeroes_at(
        &mut self, 
        offset: u64, 
        length: usize
    ) -> Result<usize, Error>;

    pub fn write_zeroes_all_at(
        &mut self, 
        offset: u64, 
        length: usize
    ) -> Result<(), Error> { ... }
}

A trait for writing zeroes to an arbitrary position in a file.

Required methods

pub fn write_zeroes_at(
&mut self,
offset: u64,
length: usize
) -> Result<usize, Error>

Write up to length bytes of zeroes starting at offset, returning how many bytes were written.

Loading content...
Provided methods

pub fn write_zeroes_all_at(
&mut self,
offset: u64,
length: usize
) -> Result<(), Error>

Write zeroes starting at offset until length bytes have been written.

This method will continuously call write_zeroes_at until the requested length is satisfied or an error is encountered.

Loading content...