Module base::wait_context

Structs

TriggeredEvent

Represents an event that has been signaled and waited for via a wait function.

WaitContext

Used to wait for multiple objects which are eligible for waiting.

Enums

EventType

Represents types of events to watch for.

Traits

EventToken

Trait that can be used to associate events with arbitrary enums when using WaitContext.

Derive Macros

EventToken

Implements the EventToken trait for a given enum.