Struct base::world::Chunk

pub struct Chunk {
    // some fields omitted
}

Represents one part of the game world.

A chunk saves CHUNK_SIZE² many hex pillars which are arranged rougly in the form of a parallelogram. See this blog post for more information (the shape is called "rhombus" there).

This type implements the Index trait and can be indexed with an AxialPoint.

Methods

impl Chunk

fn dummy() -> Self

Creates a dummy chunk for early testing. FIXME: remove

fn pillars(&self) -> &[HexPillar]