Crate base

Re-exports

pub use unix as platform;

pub use unix::net::*;

pub use unix::ioctl::*;

pub use platform::*;

Modules

common

Small system utility modules for usage by other higher level system utility modules like sys_util(on unix/linux) and win_sys_util(on windows).

descriptor_reflection

Provides infrastructure for de/serializing descriptors embedded in Rust data structures.

unix

Small system utility modules for usage by other modules.

Macros

debug

A macro for logging debug information.

error

A macro for logging an error.

generate_scoped_event

handle_eintr

Macro that retries the given expression every time its result indicates it was interrupted (i.e. returned EINTR). This is useful for operations that are prone to being interrupted by signals, such as blocking syscalls.

handle_eintr_errno

Macro that retries the given expression every time its result indicates it was interrupted. It is intended to use with system functions that signal error by returning -1 and setting errno to appropriate error code (EINTR, EINVAL, etc.) Most of standard non-reentrant libc functions use this way of signalling errors.

handle_eintr_rc

Macro that retries the given expression every time its result indicates it was interrupted. It is intended to use with system functions that return EINTR and other error codes directly as their result. Most of reentrant functions use this way of signalling errors.

info

A macro for logging info.

ioctl_expr

Raw macro to declare the expression that calculates an ioctl number

ioctl_io_nr

Declare an ioctl that transfers no data.

ioctl_ioc_nr

Raw macro to declare a function that returns an ioctl number.

ioctl_ior_nr

Declare an ioctl that reads data.

ioctl_iow_nr

Declare an ioctl that writes data.

ioctl_iowr_nr

Declare an ioctl that reads and writes data.

log

A macro for logging at an arbitrary priority level.

syscall

volatile_at_impl

volatile_impl

warn

A macro for logging a warning.

Structs

Event

See EventFd for struct- and method-level documentation.

FakeTimer

See FakeTimerFd for struct- and method-level documentation.

MemoryMapping

See MemoryMapping for struct- and method-level documentation.

MemoryMappingBuilder

RecvTube

A Tube end which can only recv messages.

ScopedEvent

An Event wrapper which triggers when it goes out of scope.

SendTube

A Tube end which can only send messages. Cloneable.

SharedMemory

See SharedMemory for struct- and method-level documentation.

Timer

See TimerFd for struct- and method-level documentation.

TriggeredEvent

Represents an event that has been signaled and waited for via a wait function.

Tube

Bidirectional tube that support both send and recv.

WaitContext

Used to wait for multiple objects which are eligible for waiting.

Enums

EventReadResult

Wrapper around the return value of doing a read on an EventFd which distinguishes between getting a valid count of the number of times the eventfd has been written to and timing out waiting for the count to be non-zero.

EventType

Represents types of events to watch for.

TubeError

Traits

AsRawDescriptors

A trait similar to AsRawDescriptor but supports an arbitrary number of descriptors.

CloseNotifier

EventToken

MemoryMappingBuilderUnix

MemoryMappingUnix

ReadNotifier

SharedMemoryUnix

Functions

ioctl⚠

Run an ioctl with no arguments.

ioctl_with_mut_ptr⚠

Run an ioctl with a mutable raw pointer.

ioctl_with_mut_ref⚠

Run an ioctl with a mutable reference.

ioctl_with_ptr⚠

Run an ioctl with a raw pointer.

ioctl_with_ref⚠

Run an ioctl with an immutable reference.

ioctl_with_val⚠

Run an ioctl with a single value argument.

validate_raw_descriptor

Verifies that |raw_descriptor| is actually owned by this process and duplicates it to ensure that we have a unique handle to it.

wrap_descriptor

Wraps an AsRawDescriptor in the simple Descriptor struct, which has AsRawFd methods for interfacing with sys_util

Type Definitions

TubeResult