Crate base

Safe, cross-platform-compatible wrappers for system interfaces.

Re-exports

pub use external_mapping::Error as ExternalMappingError;

pub use external_mapping::ExternalMapping;

pub use external_mapping::Result as ExternalMappingResult;

pub use platform::ioctl::ioctl;

pub use platform::ioctl::ioctl_with_mut_ptr;

pub use platform::ioctl::ioctl_with_mut_ref;

pub use platform::ioctl::ioctl_with_ptr;

pub use platform::ioctl::ioctl_with_ref;

pub use platform::ioctl::ioctl_with_val;

pub use platform::ioctl::IoctlNr;

pub use sys::platform;

pub use sys::unix;

pub use unix::net;

pub use platform::get_max_open_files;

pub use platform::add_fd_flags;

pub use platform::clear_fd_flags;

pub use platform::safe_descriptor_from_path;

pub use platform::validate_raw_descriptor;

pub use platform::block_signal;

pub use platform::clear_signal;

pub use platform::get_blocked_signals;

pub use platform::new_pipe_full;

pub use platform::register_rt_signal_handler;

pub use platform::signal;

pub use platform::unblock_signal;

pub use platform::Killable;

pub use platform::SIGRTMIN;

pub use platform::chown;

pub use platform::iov_max;

pub use platform::pipe;

pub use platform::flock;

pub use platform::FlockOperation;

pub use platform::getegid;

pub use platform::geteuid;

pub use platform::gettid;

pub use platform::kill_process_group;

pub use platform::reap_child;

pub use platform::net::UnixSeqpacket;

pub use platform::net::UnixSeqpacketListener;

pub use platform::net::UnlinkUnixSeqpacketListener;

pub use platform::UnlinkUnixListener;

pub use platform::deserialize_with_descriptors;

pub use platform::getpid;

pub use platform::number_of_logical_cores;

pub use platform::open_file;

pub use platform::pagesize;

pub use platform::platform_timer_resolution::enable_high_res_timers;

pub use platform::round_up_to_page_size;

pub use platform::with_as_descriptor;

pub use platform::with_raw_descriptor;

pub use platform::FileAllocate;

pub use platform::FileGetLen;

pub use platform::FileReadWriteAtVolatile;

pub use platform::FileReadWriteVolatile;

pub use platform::FileSerdeWrapper;

pub use platform::FileSetLen;

pub use platform::FileSync;

pub use platform::SerializeDescriptors;

pub use platform::UnsyncMarker;

pub use crate::descriptor::AsRawDescriptor;

pub use crate::descriptor::AsRawDescriptors;

pub use crate::descriptor::Descriptor;

pub use crate::descriptor::FromRawDescriptor;

pub use crate::descriptor::IntoRawDescriptor;

pub use crate::descriptor::SafeDescriptor;

Modules

alloc

clock

descriptor

descriptor_reflection

Provides infrastructure for de/serializing descriptors embedded in Rust data structures.

errno

event

external_mapping

mmap

notifiers

shm

sys

syslog

Facilities for sending log message to syslog.

timer

tube

wait_context

write_zeroes

Macros

debug

Logs a message at the debug level.

error

Logs a message at the error level.

handle_eintr

Macro that retries the given expression every time its result indicates it was interrupted (i.e. returned EINTR). This is useful for operations that are prone to being interrupted by signals, such as blocking syscalls.

handle_eintr_errno

Macro that retries the given expression every time its result indicates it was interrupted. It is intended to use with system functions that signal error by returning -1 and setting errno to appropriate error code (EINTR, EINVAL, etc.) Most of standard non-reentrant libc functions use this way of signalling errors.

handle_eintr_rc

Macro that retries the given expression every time its result indicates it was interrupted. It is intended to use with system functions that return EINTR and other error codes directly as their result. Most of reentrant functions use this way of signalling errors.

info

Logs a message at the info level.

ioctl_expr

Raw macro to declare the expression that calculates an ioctl number

ioctl_io_nr

Declare an ioctl that transfers no data.

ioctl_ioc_nr

Raw macro to declare a function that returns an ioctl number.

ioctl_ior_nr

Declare an ioctl that reads data.

ioctl_iow_nr

Declare an ioctl that writes data.

ioctl_iowr_nr

Declare an ioctl that reads and writes data.

syscall

trace

Logs a message at the trace level.

volatile_at_impl

volatile_impl

warn

Logs a message at the warn level.

Structs

AcpiNotifyEvent

Clock

Error

A system error In Unix systems, retrieved from errno (man 3 errno), set by a libc function that returned an error. On Windows, retrieved from GetLastError, set by a Windows function that returned an error

Event

See the platform-specific Event struct for struct- and method-level documentation.

EventContext

Used to poll multiple objects that have file descriptors.

EventFd

A safe wrapper around a Linux eventfd (man 2 eventfd).

FakeClock

A fake clock that can be used in tests to give exact control over the time. For a code example, see the tests in base/src/timer.rs.

FakeTimer

FakeTimer: For use in tests.

LayoutAllocation

A contiguous memory allocation with a specified size and alignment, with a Drop impl to perform the deallocation.

MemfdSeals

A set of memfd seals.

MemoryMapping

See MemoryMapping for struct- and method-level documentation.

MemoryMappingArena

Tracks Fixed Memory Maps within an anonymous memory-mapped fixed-sized arena in the current process.

MemoryMappingBuilder

NetlinkGenericSocket

Safe wrapper for NETLINK_GENERIC netlink sockets.

Protection

Memory access type for anonymous shared memory mapping.

RecvTube

A Tube end which can only recv messages.

SendTube

A Tube end which can only send messages. Cloneable.

SharedMemory

See SharedMemory for struct- and method-level documentation.

SignalFd

A safe wrapper around a Linux signalfd (man 2 signalfd).

StreamChannel

An abstraction over named pipes and unix socketpairs. This abstraction can be used in a blocking and non blocking mode.

Timer

TriggeredEvent

Represents an event that has been signaled and waited for via a wait function.

Tube

Bidirectional tube that support both send and recv.

WaitContext

Used to wait for multiple objects which are eligible for waiting.

Enums

BlockingMode

EventReadResult

Result of reading an Event with a timeout.

EventType

Represents types of events to watch for.

FileFlags

FramingMode

MmapError

TubeError

VmEventType

Constants

INVALID_DESCRIPTOR

SCM_SOCKET_MAX_FD_COUNT

The maximum number of FDs that can be sent in a single send.

Traits

CloseNotifier

EnabledHighResTimer

An empty trait that helps reset timer resolution to its previous state.

EventToken

Trait that can be used to associate events with arbitrary enums when using WaitContext.

MappedRegion

A range of memory that can be msynced, for abstracting over different types of memory mappings.

MemoryMappingBuilderUnix

MemoryMappingUnix

PunchHole

A trait for deallocating space in a file.

ReadNotifier

ScmSocket

Trait for file descriptors can send and receive socket control messages via sendmsg and recvmsg.

SharedMemoryUnix

Terminal

Trait for file descriptors that are TTYs, according to isatty(3).

WriteZeroesAt

A trait for writing zeroes to an arbitrary position in a file.

Functions

clear_descriptor_cloexec

Clears CLOEXEC flag on descriptor

clone_descriptor

Clones descriptor, returning a new RawDescriptor that refers to the same open file description as descriptor. The cloned descriptor will have the FD_CLOEXEC flag set but will not share any other file descriptor flags with descriptor.

drop_capabilities

Drops all capabilities (permitted, inheritable, and effective) from the current process.

enable_core_scheduling

Enable experimental core scheduling for the current thread.

errno_result

Returns the last errno as a Result that is always an error.

generate_uuid

Creates a UUID.

get_cpu_affinity

get_filesystem_type

Obtain file system type of the file system that the file is served from.

kernel_has_memfd

Checks if the kernel we are running on has memfd_create. It was introduced in 3.17. Only to be used from tests to prevent running on ancient kernels that won’t support the functionality anyways.

read_raw_stdin

Read raw bytes from stdin.

set_cpu_affinity

Set the CPU affinity of the current thread to a given set of CPUs.

set_rt_prio_limit

Enables real time thread priorities in the current thread up to limit.

set_rt_round_robin

Sets the current thread to be scheduled using the round robin real time class with priority.

Type Definitions

RawDescriptor

Result

TubeResult

Derive Macros

EventToken

Implements the EventToken trait for a given enum.