Crate base

Re-exports

pub use sys::platform;

pub use external_mapping::Error as ExternalMappingError;

pub use external_mapping::Result as ExternalMappingResult;

pub use external_mapping::*;

pub use platform::ioctl::ioctl;

pub use platform::ioctl::ioctl_with_mut_ptr;

pub use platform::ioctl::ioctl_with_mut_ref;

pub use platform::ioctl::ioctl_with_ptr;

pub use platform::ioctl::ioctl_with_ref;

pub use platform::ioctl::ioctl_with_val;

pub use platform::ioctl::*;

pub use scoped_event_macro::*;

pub use sys::unix;

pub use unix::net::*;

pub use crate::descriptor::AsRawDescriptor;

pub use crate::descriptor::AsRawDescriptors;

pub use crate::descriptor::Descriptor;

pub use crate::descriptor::FromRawDescriptor;

pub use crate::descriptor::IntoRawDescriptor;

pub use crate::descriptor::SafeDescriptor;

pub use platform::*;

Modules

alloc

clock

descriptor

descriptor_reflection

Provides infrastructure for de/serializing descriptors embedded in Rust data structures.

errno

event

external_mapping

mmap

notifiers

scoped_event_macro

shm

sys

syslog

Facilities for sending log message to syslog.

timer

tube

wait_context

Macros

debug

Logs a message at the debug level.

error

Logs a message at the error level.

generate_scoped_event

handle_eintr

Macro that retries the given expression every time its result indicates it was interrupted (i.e. returned EINTR). This is useful for operations that are prone to being interrupted by signals, such as blocking syscalls.

handle_eintr_errno

Macro that retries the given expression every time its result indicates it was interrupted. It is intended to use with system functions that signal error by returning -1 and setting errno to appropriate error code (EINTR, EINVAL, etc.) Most of standard non-reentrant libc functions use this way of signalling errors.

handle_eintr_rc

Macro that retries the given expression every time its result indicates it was interrupted. It is intended to use with system functions that return EINTR and other error codes directly as their result. Most of reentrant functions use this way of signalling errors.

info

Logs a message at the info level.

ioctl_expr

Raw macro to declare the expression that calculates an ioctl number

ioctl_io_nr

Declare an ioctl that transfers no data.

ioctl_ioc_nr

Raw macro to declare a function that returns an ioctl number.

ioctl_ior_nr

Declare an ioctl that reads data.

ioctl_iow_nr

Declare an ioctl that writes data.

ioctl_iowr_nr

Declare an ioctl that reads and writes data.

log

The standard logging macro.

log_enabled

Determines if a message logged at the specified level in that module will be logged.

syscall

trace

Logs a message at the trace level.

volatile_at_impl

volatile_impl

warn

Logs a message at the warn level.

Structs

Clock

Error

A system error In Unix systems, retrieved from errno (man 3 errno), set by a libc function that returned an error. On Windows, retrieved from GetLastError, set by a Windows function that returned an error

Event

See PlatformEvent for struct- and method-level documentation.

FakeClock

A fake clock that can be used in tests to give exact control over the time. For a code example, see the tests in base/src/timer.rs.

FakeTimer

FakeTimer: For use in tests.

LayoutAllocation

A contiguous memory allocation with a specified size and alignment, with a Drop impl to perform the deallocation.

MemoryMapping

See MemoryMapping for struct- and method-level documentation.

MemoryMappingBuilder

Metadata

Metadata about a log message.

MetadataBuilder

Builder for Metadata.

ParseLevelError

The type returned by from_str when the string doesn’t match any of the log levels.

Record

The “payload” of a log message.

RecordBuilder

Builder for Record.

RecvTube

A Tube end which can only recv messages.

ScopedEvent

An Event wrapper which triggers when it goes out of scope.

SendTube

A Tube end which can only send messages. Cloneable.

SetLoggerError

The type returned by set_logger if set_logger has already been called.

SharedMemory

See SharedMemory for struct- and method-level documentation.

Timer

TriggeredEvent

Represents an event that has been signaled and waited for via a wait function.

Tube

Bidirectional tube that support both send and recv.

WaitContext

Used to wait for multiple objects which are eligible for waiting.

Enums

EventReadResult

Wrapper around the return value of doing a read on an EventFd which distinguishes between getting a valid count of the number of times the eventfd has been written to and timing out waiting for the count to be non-zero.

EventType

Represents types of events to watch for.

Level

An enum representing the available verbosity levels of the logger.

LevelFilter

An enum representing the available verbosity level filters of the logger.

TubeError

Constants

STATIC_MAX_LEVEL

The statically resolved maximum log level.

Traits

CloseNotifier

EventToken

Log

A trait encapsulating the operations required of a logger.

MemoryMappingBuilderUnix

MemoryMappingUnix

ReadNotifier

Functions

errno_result

Returns the last errno as a Result that is always an error.

logger

Returns a reference to the logger.

max_level

Returns the current maximum log level.

set_logger

Sets the global logger to a &'static Log.

set_logger_racy⚠

A thread-unsafe version of set_logger.

set_max_level

Sets the global maximum log level.

Type Definitions

Result

TubeResult