Crate base

Functionality used by both the plantex server and the plantex client.

the math module reexports everything from the cgmath crate and defines a few own type

the world module is all about saving and managing the game world

Reexports

pub extern crate rand;

pub extern crate noise;

Modules

gen
Functionality about procedurally generating content.

math
This module reexports everything from the cgmath crate and defines a few own types, such as AxialVector and AxialPoint.

msg
Messages for client-server-communication.

prop
Contains types of the available props.

world
Types and constants to represent a game world.