Crate base

A crate to hold very common types in Servo.

You should almost never need to add a data type to this crate. Instead look for a more shared crate that has fewer dependents.

Modules

cross_process_instant

An implementation of a monotonic, nanosecond precision timer, like std::time::Instant that can be serialized and compared across processes.

generic_channel

Enum wrappers to be able to select different channel implementations at runtime.

id

Namespaces and ids shared by many crates in Servo.

print_tree

text

unicode_block 🔒

Structs

Epoch

A struct for denoting the age of messages; prevents race conditions.

Traits

WebRenderEpochToU16