Struct base::MemoryMapping

pub struct MemoryMapping {
    pub(crate) mapping: PlatformMmap,
    pub(crate) _file_descriptor: Option<SafeDescriptor>,
}

See MemoryMapping for struct- and method-level documentation.

Fields
mapping: PlatformMmap_file_descriptor: Option<SafeDescriptor>
Implementations

impl MemoryMapping

pub fn write_slice(&self, buf: &[u8], offset: usize) -> Result<usize, Error>

pub fn read_slice(&self, buf: &mut [u8], offset: usize) -> Result<usize, Error>

pub fn write_obj<T: DataInit>(&self, val: T, offset: usize) -> Result<(), Error>

Writes an object to the memory region at the specified offset. Returns Ok(()) if the object fits, or Err if it extends past the end.

This method is for writing to regular memory. If writing to a mapped I/O region, use MemoryMapping::write_obj_volatile.

Examples

Write a u64 at offset 16.

    let res = mem_map.write_obj(55u64, 16);
    assert!(res.is_ok());

pub fn read_obj<T: DataInit>(&self, offset: usize) -> Result<T, Error>

Reads on object from the memory region at the given offset. Reading from a volatile area isn’t strictly safe as it could change mid-read. However, as long as the type T is plain old data and can handle random initialization, everything will be OK.

This method is for reading from regular memory. If reading from a mapped I/O region, use MemoryMapping::read_obj_volatile.

Examples

Read a u64 written to offset 32.

    let res = mem_map.write_obj(55u64, 32);
    assert!(res.is_ok());
    let num: u64 = mem_map.read_obj(32).unwrap();
    assert_eq!(55, num);

pub fn write_obj_volatile<T: DataInit>(
&self,
val: T,
offset: usize
) -> Result<(), Error>

Writes an object to the memory region at the specified offset. Returns Ok(()) if the object fits, or Err if it extends past the end.

The write operation will be volatile, i.e. it will not be reordered by the compiler and is suitable for I/O, but must be aligned. When writing to regular memory, prefer MemoryMapping::write_obj.

Examples

Write a u32 at offset 16.

    let res = mem_map.write_obj_volatile(0xf00u32, 16);
    assert!(res.is_ok());

pub fn read_obj_volatile<T: DataInit>(&self, offset: usize) -> Result<T, Error>

Reads on object from the memory region at the given offset. Reading from a volatile area isn’t strictly safe as it could change mid-read. However, as long as the type T is plain old data and can handle random initialization, everything will be OK.

The read operation will be volatile, i.e. it will not be reordered by the compiler and is suitable for I/O, but must be aligned. When reading from regular memory, prefer MemoryMapping::read_obj.

Examples

Read a u32 written to offset 16.

    let res = mem_map.write_obj(0xf00u32, 16);
    assert!(res.is_ok());
    let num: u32 = mem_map.read_obj(16).unwrap();
    assert_eq!(0xf00, num);

pub fn msync(&self) -> Result<(), Error>

impl CrateMemoryMapping

pub fn use_hugepages(&self) -> Result<(), Error>

pub fn read_to_memory(
&self,
mem_offset: usize,
src: &dyn AsRawDescriptor,
count: usize
) -> Result<(), Error>

pub fn write_from_memory(
&self,
mem_offset: usize,
dst: &dyn AsRawDescriptor,
count: usize
) -> Result<(), Error>

pub fn from_raw_ptr(
addr: RawDescriptor,
size: usize
) -> Result<CrateMemoryMapping, Error>